--- a/src/cli/commands/new.ts
+++ b/src/cli/commands/new.ts
@@ -1270,7 +1270,7 @@ async function createPackageJson(
       preview: '0x1 preview'
     },
     dependencies: {
-      "0x1": '^0.0.281' // Use current version with caret for compatibility
+      "0x1": '^0.0.282' // Use current version with caret for compatibility
     },
     devDependencies: {
       typescript: '^5.4.5'